simple-web-app/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
dashscope>=1.17.0
python-dotenv>=1.0.0
python-socketio>=5.8.0
eventlet>=0.33.0
requests>=2.31.0